+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - changed
   - Parse `ComplexType`s recursively
   - Rename `InstanceParser::parse_instance` to `InstanceParser::parse`
+  - Rename `LinkbaseParser::parse_linkbase` to `InstanceParser::parse`
 - removed
   - Remove `tuple_children` and `compositor` from `Concept`
   - Remove `context` for `XbrlError::FileOpen` and `XbrlError::FileRead`